Rename module to sessions
