--- libs/service-portal/activities/src/index.ts
@@ -1,2 +1,0 @@
-export { activitiesModule } from './module'
-export { activitiesNavigation } from './lib/navigation'
