--- libs/service-portal/activities/src/lib/messages.ts
@@ -1,21 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export const m = defineMessages({
-  activities: {
-    id: 'portals-my-pages.activities:name',
-    defaultMessage: 'Hreyfingar',
-  },
-  activitiesDescription: {
-    id: 'portals-my-pages.activities:description',
-    defaultMessage: 'Yfirlit yfir innskráningar hreyfingar á Ísland.is',
-  },
-  sessions: {
-    id: 'portals-my-pages.activities:sessions',
-    defaultMessage: 'Innskráningar',
-  },
-  sessionsHeaderIntro: {
-    id: 'portals-my-pages.activities:sessions-header-intro',
-    defaultMessage:
-      'Hér er yfirlit yfir innskráningar þínar á Ísland.is, sem þú sjálfur, aðrir hafa gert í þínu umboði eða sem þú hefur gert í umboði fyrir aðra.',
-  },
-})
