+++ libs/service-portal/sessions/src/index.ts
@@ -1,0 +1,2 @@
+export { sessionsModule } from './module'
+export { sessionsNavigation } from './lib/navigation'
