+++ libs/service-portal/sessions/src/lib/messages.ts
@@ -1,0 +1,17 @@
+import { defineMessages } from 'react-intl'
+
+export const m = defineMessages({
+  sessions: {
+    id: 'portals-my-pages.session-history:sessions',
+    defaultMessage: 'Innskráningar',
+  },
+  sessionsDescription: {
+    id: 'portals-my-pages.session-history:description',
+    defaultMessage: 'Yfirlit yfir innskráningar hreyfingar á Ísland.is',
+  },
+  sessionsHeaderIntro: {
+    id: 'portals-my-pages.session-history:sessions-header-intro',
+    defaultMessage:
+      'Hér er yfirlit yfir innskráningar þínar á Ísland.is, sem þú sjálfur, aðrir hafa gert í þínu umboði eða sem þú hefur gert í umboði fyrir aðra.',
+  },
+})
