+++ libs/service-portal/sessions/src/lib/paths.ts
@@ -1,3 +1,3 @@
-export enum ActivitiesPaths {
+export enum SessionsPaths {
   Sessions = '/adgangsstyring/innskraningar',
 }
