+++ tsconfig.base.json
@@ -596,8 +596,8 @@
       "@island.is/regulations": ["libs/regulations/src/index.ts"],
       "@island.is/regulations/*": ["libs/regulations/src/sub/*"],
       "@island.is/residence-history": ["libs/residence-history/src/index.ts"],
-      "@island.is/service-portal/activities": [
-        "libs/service-portal/activities/src/index.ts"
+      "@island.is/service-portal/sessions": [
+        "libs/service-portal/sessions/src/index.ts"
       ],
       "@island.is/service-portal/applications": [
         "libs/service-portal/applications/src/index.ts"
