+++ workspace.json
@@ -245,7 +245,6 @@
     "residence-history": "libs/residence-history",
     "samradsgatt": "apps/samradsgatt",
     "service-portal": "apps/service-portal",
-    "service-portal-activities": "libs/service-portal/activities",
     "service-portal-applications": "libs/service-portal/applications",
     "service-portal-assets": "libs/service-portal/assets",
     "service-portal-constants": "libs/service-portal/constants",
@@ -265,6 +264,7 @@
     "service-portal-icelandic-names-registry": "libs/service-portal/icelandic-names-registry",
     "service-portal-information": "libs/service-portal/information",
     "service-portal-licenses": "libs/service-portal/licenses",
+    "service-portal-sessions": "libs/service-portal/sessions",
     "service-portal-settings-personal-information": "libs/service-portal/settings/personal-information",
     "service-portal-vehicles": "libs/service-portal/vehicles",
     "services-auth-admin-api": "apps/services/auth/admin-api",
